Decrease delay
Delay is only added to prevent buffer-overflowing the microcontroller when sending lots of data (like a system backup). Too much delay will cause the calculator to time out while uploading a backup/program, so we want to keep this value shorter than the real amount of time that a byte takes. Too high of a delay is more likely to cause problems than too low, so I'm playing it safe and keeping it at 1ms.